chat-inference/requirements.txt
# Chat Inference: Dify-free classifier + RAG + LLM pipeline
# Uses same Qdrant (Dify's DB) via RAG Backend API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
pydantic>=2.0
pydantic-settings>=2.0
pyjwt>=2.8.0
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0
jinja2>=3.0.0
minio>=7.2.0

# LangChain + Gemini (same as RAG Backend / Dify)
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
